Cleaning up exitable vector value impls
